feat(desktop): 抽取 provider 连接对话框并在 OOBE 连接步骤复用
- 从 models-settings-panel 抽出 provider-connect-dialog 与 provider-picker-rows
- OOBE Step 3 连接提供商与完成落盘，Done 写入 onboardingCompleted
- Step 3 离场时冻结列表底缘渐隐（mask-styles）
