+++ apps/desktop/src/components/settings/models/models-settings-panel.tsx
@@ -1,18 +1,8 @@
-import { useCallback, useEffect, useMemo, useRef, useState, type KeyboardEvent } from "react";
+import { useCallback, useMemo, useRef, useState, type KeyboardEvent } from "react";
 import { useTranslation } from "react-i18next";
-import { LoaderCircle, RefreshCw } from "lucide-react";
+import { LoaderCircle } from "lucide-react";
 
 import { ModelCatalogDetailPanel } from "@/components/model-catalog-detail-panel";
-import { ProviderIcon } from "@/components/provider-icon";
-import {
-  connectTransportOptionSummary,
-  connectTransportOptionsForProvider,
-  defaultConnectTransportKind,
-  providerShowsConnectTransportPicker,
-  resolveConnectTransportKindForProvider,
-  resolveCustomConnectApiBase,
-} from "@/components/settings/models/connect-transport";
-import { ModelCapabilitiesCombobox } from "@/components/settings/models/model-capabilities-combobox";
 import {
   type ModelDefaultAssignments,
   type ModelDefaultRole,
@@ -21,15 +11,20 @@ import {
   canAssignAsImageGenerationModel,
   canAssignAsLightweightChatModel,
   canAssignAsVideoGenerationModel,
-  defaultCustomModelCapabilities,
   getSupportedModelDefaultRoles,
   modelDefaultActionLabel,
   modelRefKey,
-  normalizeModelCapabilitySelection,
   resolveEffectiveModelRef,
   settingsModelRef,
 } from "@/components/settings/models/model-defaults";
 import { ModelSettingsRowButton, ModelSettingsRowWithHover } from "@/components/settings/models/model-settings-row";
+import { ProviderConnectDialog } from "@/components/settings/models/provider-connect-dialog";
+import {
+  ProviderPickerRowButton,
+  filterProviderRows,
+  localizedProviderRows,
+  providerPickerLabel,
+} from "@/components/settings/models/provider-picker-rows";
 import type { SettingsFormState, SettingsViewProps } from "@/components/settings/types";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
@@ -43,73 +38,20 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import { DesktopFormInput, DesktopFormTextarea } from "@/components/ui/desktop-form-field";
+import { DesktopFormInput } from "@/components/ui/desktop-form-field";
 import { Label } from "@/components/ui/label";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 import { Tooltip, TooltipContent } from "@/components/ui/tooltip";
 import { ScrollArea } from "@/components/ui/scroll-area";
-import {
-  hasBedrockIamCredentials,
-  hasGoogleVertexServiceAccountCredentials,
-} from "@/lib/provider-runtime-credentials";
-import {
-  PROVIDER_PICKER_ROWS,
-  defaultProviderConnectSite,
-  listProviderConnectSiteOptions,
-  providerConnectSiteRequiresWorkspaceId,
-  providerSupportsSiteSelection,
-  resolveProviderConnectApiBase,
-} from "@spiritagent/host-internal/model-provider-presets";
 import {
   buildModelCatalogDetailMap,
   buildModelCatalogDisplayTitleMap,
   modelCatalogScopeEntryKey,
   modelDisplayTitleFromMap,
-  modelSettingsRowAriaLabel,
 } from "@/lib/model-catalog-detail";
-import { modelCapabilityLabel } from "@/lib/model-capability-label";
-import {
-  DESKTOP_FORM_FIELD_TRIGGER_INNER,
-  DESKTOP_FORM_INPUT_SHELL,
-} from "@/lib/desktop-chrome";
-import { parseModelContextLength } from "@/lib/model-context-length";
 import { runAfterRadixOverlayClose } from "@/lib/overlay-motion";
-import { cn } from "@/lib/utils";
-import type {
-  AddModelRequest,
-  AddProviderModelsRequest,
-  DesktopModelCapability,
-  DesktopModelProvider,
-  DesktopTransportKind,
-  PreviewModelCatalogEntry,
-  PreviewModelsRequest,
-  PreviewModelsResponse,
-  ProviderGroupV2,
-} from "@/types";
-import { bedrockApiBaseFromRegion } from "@spiritagent/host-internal/bedrock-region";
-import {
-  defaultPresetProviderGroupId,
-  modelRefsEqual,
-  slugifyProviderGroupLabel,
-} from "@spiritagent/host-internal/config-v2";
-import {
-  bedrockMantleApiBaseFromRegion,
-  isBedrockMantleOpenAiModel,
-} from "@spiritagent/host-internal/bedrock-mantle";
-import { azureApiBaseFromResourceName, isValidAzureResourceName } from "@spiritagent/host-internal/azure-resource";
-import {
-  cloudflareAiGatewayApiBaseFromAccountId,
-  isValidCloudflareAccountId,
-  isValidCloudflareGatewayId,
-} from "@spiritagent/host-internal/cloudflare-ai-gateway-resource";
-import { vertexApiBaseFromProjectAndLocation } from "@spiritagent/host-internal/google-vertex-endpoints";
-import {DESKTOP_EDITOR_TAB_CLASS, DESKTOP_LIST_ITEM_PRIMARY_CLASS, DESKTOP_PAGE_TITLE_CLASS } from "@/lib/desktop-typography";
+import type { DesktopModelProvider, ProviderGroupV2 } from "@/types";
+import { modelRefsEqual } from "@spiritagent/host-internal/config-v2";
+import { DESKTOP_LIST_ITEM_PRIMARY_CLASS, DESKTOP_PAGE_TITLE_CLASS } from "@/lib/desktop-typography";
 
 export function ModelsSettingsPanel({
   settings,
@@ -141,41 +83,11 @@ export function ModelsSettingsPanel({
   const providerSearchInputRef = useRef<HTMLInputElement>(null);
   const providerListRef = useRef<HTMLDivElement>(null);
   const [connectDialogOpen, setConnectDialogOpen] = useState(false);
-  const [selectedProvider, setSelectedProvider] = useState<DesktopModelProvider | null>(null);
-  const [connectApiKey, setConnectApiKey] = useState("");
-  const [connectAwsRegion, setConnectAwsRegion] = useState("");
-  const [connectAzureResourceName, setConnectAzureResourceName] = useState("");
-  const [connectCloudflareAccountId, setConnectCloudflareAccountId] = useState("");
-  const [connectCloudflareGatewayId, setConnectCloudflareGatewayId] = useState("");
-  const [connectAccessKeyId, setConnectAccessKeyId] = useState("");
-  const [connectSecretAccessKey, setConnectSecretAccessKey] = useState("");
-  const [connectVertexProject, setConnectVertexProject] = useState("");
-  const [connectVertexLocation, setConnectVertexLocation] = useState("");
-  const [connectVertexClientEmail, setConnectVertexClientEmail] = useState("");
-  const [connectVertexPrivateKey, setConnectVertexPrivateKey] = useState("");
-  const [connectName, setConnectName] = useState("");
-  const [connectApiBase, setConnectApiBase] = useState("");
-  const [connectCapabilities, setConnectCapabilities] = useState<DesktopModelCapability[]>(
-    defaultCustomModelCapabilities,
-  );
-  const [connectContextLength, setConnectContextLength] = useState("");
-  const [connectTransportKind, setConnectTransportKind] = useState<DesktopTransportKind>(
-    "openai-compatible",
-  );
-  const [connectProviderSite, setConnectProviderSite] = useState("");
-  const [connectAlibabaBillingMode, setConnectAlibabaBillingMode] = useState<
-    "standard" | "token-plan"
-  >("standard");
-  const [connectStepfunBillingMode, setConnectStepfunBillingMode] = useState<
-    "standard" | "step-plan"
-  >("standard");
-  const [connectAlibabaWorkspaceId, setConnectAlibabaWorkspaceId] = useState("");
-  const [customConnectMode, setCustomConnectMode] = useState<"single" | "bulk">(
-    "single",
-  );
-  const [bedrockConnectMode, setBedrockConnectMode] = useState<"bearer" | "iam">("bearer");
-  const [vertexConnectMode, setVertexConnectMode] = useState<"adc" | "service-account" | "express">("adc");
-  const [connectCustomGroupLabel, setConnectCustomGroupLabel] = useState("");
+  /** session 每次发起连接自增，作为 ProviderConnectDialog 的 key 重挂载以重置表单。 */
+  const [connectTarget, setConnectTarget] = useState<{
+    provider: DesktopModelProvider;
+    session: number;
+  } | null>(null);
   const [modelDefaultsDialogTarget, setModelDefaultsDialogTarget] = useState<string | null>(null);
   const [modelDefaultsDialogOpen, setModelDefaultsDialogOpen] = useState(false);
   const [modelDefaultAssignments, setModelDefaultAssignments] = useState<ModelDefaultAssignments>({
@@ -276,90 +188,24 @@ export function ModelsSettingsPanel({
       .filter((model): model is SettingsModelProfile => model !== undefined);
   }
 
-  function resolveConnectGroupId(provider: DesktopModelProvider): string {
-    if (provider === "custom") {
-      const label = connectCustomGroupLabel.trim();
-      return label ? slugifyProviderGroupLabel(label) : defaultPresetProviderGroupId("custom");
-    }
-    return defaultPresetProviderGroupId(provider);
-  }
   function providerLabel(provider: DesktopModelProvider): string {
-    const row = PROVIDER_PICKER_ROWS.find((item) => item.id === provider);
-    return row ? String(t(row.labelKey, { defaultValue: row.fallbackLabel })) : provider;
+    return providerPickerLabel((key, options) => String(t(key, options)), provider);
   }
-
-  const resetConnectWizard = () => {
-    setConnectApiKey("");
-    setConnectAwsRegion("");
-    setConnectAzureResourceName("");
-    setConnectCloudflareAccountId("");
-    setConnectCloudflareGatewayId("");
-    setConnectAccessKeyId("");
-    setConnectSecretAccessKey("");
-    setConnectVertexProject("");
-    setConnectVertexLocation("");
-    setConnectVertexClientEmail("");
-    setConnectVertexPrivateKey("");
-    setConnectName("");
-    setConnectApiBase("");
-    setConnectCapabilities(defaultCustomModelCapabilities);
-    setConnectContextLength("");
-    setConnectTransportKind("openai-compatible");
-    setConnectProviderSite("");
-    setConnectAlibabaBillingMode("standard");
-    setConnectStepfunBillingMode("standard");
-    setConnectAlibabaWorkspaceId("");
-    setConnectCustomGroupLabel("");
-    setCustomConnectMode("single");
-    setBedrockConnectMode("bearer");
-    setVertexConnectMode("adc");
-    setSelectedProvider(null);
-  };
-
-  const resetConnectTransportKindForProvider = (provider: DesktopModelProvider) => {
-    setConnectTransportKind(defaultConnectTransportKind(provider));
-    setConnectProviderSite(defaultProviderConnectSite(provider) ?? "");
-  };
 
   const openProviderPicker = () => {
-    resetConnectWizard();
     setProviderQuery("");
     setProviderDialogOpen(true);
   };
 
   const startConnect = (id: DesktopModelProvider) => {
     setProviderDialogOpen(false);
-    setSelectedProvider(id);
-    setConnectApiKey("");
-    setConnectAwsRegion("");
-    setConnectAzureResourceName("");
-    setConnectCloudflareAccountId("");
-    setConnectCloudflareGatewayId("");
-    setConnectAccessKeyId("");
-    setConnectSecretAccessKey("");
-    setConnectVertexProject("");
-    setConnectVertexLocation("");
-    setConnectVertexClientEmail("");
-    setConnectVertexPrivateKey("");
-    setConnectName("");
-    setConnectApiBase("");
-    setConnectCapabilities(defaultCustomModelCapabilities);
-    resetConnectTransportKindForProvider(id);
-    setConnectAlibabaBillingMode("standard");
-    setConnectStepfunBillingMode("standard");
-    setConnectCustomGroupLabel("");
-    setCustomConnectMode("single");
-    setBedrockConnectMode("bearer");
-    setVertexConnectMode("adc");
+    setConnectTarget((prev) => ({ provider: id, session: (prev?.session ?? 0) + 1 }));
     setConnectDialogOpen(true);
   };
 
-  const localizedProviderRows = PROVIDER_PICKER_ROWS.map((row) => ({
-    ...row,
-    label: providerLabel(row.id),
-  }));
-  const filteredProviders = localizedProviderRows.filter((row) =>
-    row.label.toLowerCase().includes(providerQuery.trim().toLowerCase()),
+  const filteredProviders = filterProviderRows(
+    localizedProviderRows((key, options) => String(t(key, options))),
+    providerQuery,
   );
 
   const collectProviderListButtons = (container: HTMLElement | null): HTMLElement[] => {
@@ -411,11 +257,6 @@ export function ModelsSettingsPanel({
       }
     }
   };
-
-  const selectedProviderLabel =
-    selectedProvider === null
-      ? t('settings.connectProvider')
-      : providerLabel(selectedProvider);
 
   const openModelDefaultsDialog = (model: SettingsModelProfile) => {
     const modelRef = settingsModelRef(model);
@@ -567,387 +408,6 @@ export function ModelsSettingsPanel({
     }
 
     openModelDefaultsDialog(model);
-  };
-
-  const connectTransportKindForRequest = resolveConnectTransportKindForProvider(
-    selectedProvider,
-    connectTransportKind,
-  );
-
-  const connectAlibabaIsTokenPlan =
-    selectedProvider === "alibaba" && connectAlibabaBillingMode === "token-plan";
-  const connectAlibabaBillingModeForRequest = connectAlibabaIsTokenPlan
-    ? ("token-plan" as const)
-    : undefined;
-  const connectStepfunIsStepPlan =
-    selectedProvider === "stepfun" && connectStepfunBillingMode === "step-plan";
-  const connectStepfunBillingModeForRequest = connectStepfunIsStepPlan
-    ? ("step-plan" as const)
-    : undefined;
-  const connectProviderSiteForRequest = connectAlibabaIsTokenPlan
-    ? undefined
-    : connectProviderSite.trim() || undefined;
-  const connectAlibabaWorkspaceIdForRequest = connectAlibabaIsTokenPlan
-    ? undefined
-    : connectAlibabaWorkspaceId.trim() || undefined;
-  const connectAlibabaWorkspaceRequired =
-    selectedProvider === "alibaba"
-    && !connectAlibabaIsTokenPlan
-    && connectProviderSiteForRequest !== undefined
-    && providerConnectSiteRequiresWorkspaceId("alibaba", connectProviderSiteForRequest);
-  const connectProviderSiteForApiBase =
-    connectProviderSiteForRequest
-    && selectedProvider !== null
-    && providerConnectSiteRequiresWorkspaceId(selectedProvider, connectProviderSiteForRequest)
-    && !connectAlibabaWorkspaceIdForRequest
-      ? undefined
-      : connectProviderSiteForRequest;
-
-  const effectiveApiBase =
-    selectedProvider === null
-      ? ""
-      : selectedProvider === "amazon-bedrock"
-        ? bedrockApiBaseFromRegion(connectAwsRegion)
-        : selectedProvider === "azure"
-          ? azureApiBaseFromResourceName(connectAzureResourceName)
-        : selectedProvider === "cloudflare-ai-gateway"
-          ? cloudflareAiGatewayApiBaseFromAccountId(connectCloudflareAccountId)
-        : selectedProvider === "google-vertex-ai" && vertexConnectMode !== "express"
-          ? vertexApiBaseFromProjectAndLocation(connectVertexProject, connectVertexLocation)
-        : selectedProvider === "custom"
-        ? resolveCustomConnectApiBase(connectApiBase)
-        : resolveProviderConnectApiBase(
-            selectedProvider,
-            connectTransportKindForRequest ?? defaultConnectTransportKind(selectedProvider),
-            {
-              ...(connectAlibabaBillingModeForRequest
-                ? { billingMode: connectAlibabaBillingModeForRequest }
-                : {}),
-              ...(connectStepfunBillingModeForRequest
-                ? { stepfunBillingMode: connectStepfunBillingModeForRequest }
-                : {}),
-              ...(connectProviderSiteForApiBase ? { site: connectProviderSiteForApiBase } : {}),
-              ...(connectAlibabaWorkspaceIdForRequest
-                ? { workspaceId: connectAlibabaWorkspaceIdForRequest }
-                : {}),
-            },
-          );
-
-  const hasBedrockCatalogCredentials = hasBedrockIamCredentials({
-    accessKeyId: connectAccessKeyId,
-    secretAccessKey: connectSecretAccessKey,
-  });
-  const hasVertexCatalogCredentials = hasGoogleVertexServiceAccountCredentials({
-    clientEmail: connectVertexClientEmail,
-    privateKey: connectVertexPrivateKey,
-  });
-  const vertexConnectFields = {
-    ...(connectVertexProject.trim() ? { vertexProject: connectVertexProject.trim() } : {}),
-    ...(connectVertexLocation.trim() ? { vertexLocation: connectVertexLocation.trim() } : {}),
-    ...(connectVertexClientEmail.trim() ? { vertexClientEmail: connectVertexClientEmail.trim() } : {}),
-    ...(connectVertexPrivateKey.trim() ? { vertexPrivateKey: connectVertexPrivateKey.trim() } : {}),
-  };
-
-  const syncCatalogFromUpstream = async (forceRefresh: boolean) => {
-    if (selectedProvider === null) {
-      return;
-    }
-    if (selectedProvider === "amazon-bedrock") {
-      if (!connectAwsRegion.trim()) {
-        throw new Error(t('settings.bedrockRegionRequired'));
-      }
-      if (!hasBedrockCatalogCredentials) {
-        throw new Error(t('settings.bedrockCatalogIamRequired'));
-      }
-    } else if (selectedProvider === "google-vertex-ai") {
-      if (!connectVertexProject.trim() || !connectVertexLocation.trim()) {
-        throw new Error(t('settings.vertexProjectLocationRequired'));
-      }
-      if (vertexConnectMode === "service-account" && !hasVertexCatalogCredentials) {
-        throw new Error(t('settings.vertexCatalogServiceAccountRequired'));
-      }
-      if (vertexConnectMode === "express") {
-        throw new Error(t('settings.vertexExpressCatalogUnsupported'));
-      }
-    } else if (selectedProvider === "custom") {
-      if (!connectApiBase.trim()) {
-        throw new Error(t('settings.endpointRequired'));
-      }
-    } else if (!connectApiKey.trim()) {
-      throw new Error(t('settings.apiKeyRequired'));
-    }
-    if (connectAlibabaWorkspaceRequired && !connectAlibabaWorkspaceIdForRequest) {
-      throw new Error(t('settings.alibabaWorkspaceIdRequired'));
-    }
-    const res = await onPreviewModels({
-      apiBase: effectiveApiBase,
-      apiKey: connectApiKey,
-      provider: selectedProvider,
-      ...(selectedProvider === "amazon-bedrock"
-        ? {
-            awsRegion: connectAwsRegion.trim(),
-            accessKeyId: connectAccessKeyId.trim(),
-            secretAccessKey: connectSecretAccessKey.trim(),
-          }
-        : {}),
-      ...(selectedProvider === "google-vertex-ai" ? vertexConnectFields : {}),
-      ...(connectTransportKindForRequest !== undefined
-        ? { transportKind: connectTransportKindForRequest }
-        : {}),
-      ...(connectProviderSiteForRequest ? { providerSite: connectProviderSiteForRequest } : {}),
-      ...(connectAlibabaWorkspaceIdForRequest
-        ? { alibabaWorkspaceId: connectAlibabaWorkspaceIdForRequest }
-        : {}),
-      ...(connectAlibabaBillingModeForRequest
-        ? { alibabaBillingMode: connectAlibabaBillingModeForRequest }
-        : {}),
-      ...(connectStepfunBillingModeForRequest
-        ? { stepfunBillingMode: connectStepfunBillingModeForRequest }
-        : {}),
-      forceRefresh,
-    });
-    if (res.modelIds.length === 0) {
-      throw new Error(t('settings.noModelsReturned'));
-    }
-    const bulk: AddProviderModelsRequest = {
-      groupId: resolveConnectGroupId(selectedProvider),
-      apiBase: effectiveApiBase,
-      apiKey: connectApiKey,
-      modelIds: res.modelIds,
-      ...(res.models ? { modelCatalog: res.models } : {}),
-      provider: selectedProvider,
-      ...(selectedProvider === "amazon-bedrock"
-        ? {
-            awsRegion: connectAwsRegion.trim(),
-            accessKeyId: connectAccessKeyId.trim(),
-            secretAccessKey: connectSecretAccessKey.trim(),
-          }
-        : {}),
-      ...(selectedProvider === "google-vertex-ai" ? vertexConnectFields : {}),
-      ...(connectTransportKindForRequest !== undefined
-        ? { transportKind: connectTransportKindForRequest }
-        : {}),
-      ...(connectProviderSiteForRequest ? { providerSite: connectProviderSiteForRequest } : {}),
-      ...(connectAlibabaWorkspaceIdForRequest
-        ? { alibabaWorkspaceId: connectAlibabaWorkspaceIdForRequest }
-        : {}),
-      ...(connectAlibabaBillingModeForRequest
-        ? { alibabaBillingMode: connectAlibabaBillingModeForRequest }
-        : {}),
-      ...(connectStepfunBillingModeForRequest
-        ? { stepfunBillingMode: connectStepfunBillingModeForRequest }
-        : {}),
-    };
-    await onAddProviderModels(bulk);
-    setConnectDialogOpen(false);
-    runAfterRadixOverlayClose(resetConnectWizard);
-  };
-
-  const saveCustomSingle = async () => {
-    if (selectedProvider === null || selectedProvider !== "custom") {
-      return;
-    }
-    const name = connectName.trim();
-    const apiBase = effectiveApiBase;
-    if (!name) {
-      throw new Error(t('settings.modelNameRequired'));
-    }
-    if (!connectApiBase.trim()) {
-      throw new Error(t('settings.endpointRequired'));
-    }
-    if (!connectApiKey.trim()) {
-      throw new Error(t('settings.apiKeyRequired'));
-    }
-    const contextLengthRaw = connectContextLength.trim();
-    let contextLength: number | undefined;
-    if (contextLengthRaw) {
-      const parsed = parseModelContextLength(Number(contextLengthRaw));
-      if (parsed === undefined) {
-        throw new Error(t('settings.contextLengthInvalid'));
-      }
-      contextLength = parsed;
-    }
-    await onAddModel({
-      groupId: resolveConnectGroupId("custom"),
-      name,
-      apiBase,
-      apiKey: connectApiKey,
-      provider: "custom",
-      transportKind: connectTransportKind,
-      capabilities: normalizeModelCapabilitySelection(connectCapabilities),
-      ...(connectCustomGroupLabel.trim()
-        ? { customGroupLabel: connectCustomGroupLabel.trim() }
-        : {}),
-      ...(contextLength !== undefined ? { contextLength } : {}),
-    });
-    setConnectDialogOpen(false);
-    runAfterRadixOverlayClose(resetConnectWizard);
-  };
-
-  const saveBedrockBearerSingle = async () => {
-    if (selectedProvider !== "amazon-bedrock") {
-      return;
-    }
-    const awsRegion = connectAwsRegion.trim();
-    const name = connectName.trim();
-    if (!awsRegion) {
-      throw new Error(t('settings.bedrockRegionRequired'));
-    }
-    if (!name) {
-      throw new Error(t('settings.modelNameRequired'));
-    }
-    if (!connectApiKey.trim()) {
-      throw new Error(t('settings.apiKeyRequired'));
-    }
-    const contextLengthRaw = connectContextLength.trim();
-    let contextLength: number | undefined;
-    if (contextLengthRaw) {
-      const parsed = parseModelContextLength(Number(contextLengthRaw));
-      if (parsed === undefined) {
-        throw new Error(t('settings.contextLengthInvalid'));
-      }
-      contextLength = parsed;
-    }
-    await onAddModel({
-      groupId: defaultPresetProviderGroupId("amazon-bedrock"),
-      name,
-      apiBase: isBedrockMantleOpenAiModel(name)
-        ? bedrockMantleApiBaseFromRegion(awsRegion)
-        : bedrockApiBaseFromRegion(awsRegion),
-      apiKey: connectApiKey,
-      provider: "amazon-bedrock",
-      transportKind: "bedrock",
-      awsRegion,
-      ...(contextLength !== undefined ? { contextLength } : {}),
-    });
-    setConnectDialogOpen(false);
-    runAfterRadixOverlayClose(resetConnectWizard);
-  };
-
-  const saveAzureSingle = async () => {
-    if (selectedProvider !== "azure") {
-      return;
-    }
-    const azureResourceName = connectAzureResourceName.trim();
-    const name = connectName.trim();
-    if (!azureResourceName) {
-      throw new Error(t('settings.azureResourceNameRequired'));
-    }
-    if (!isValidAzureResourceName(azureResourceName)) {
-      throw new Error(t('settings.azureResourceNameInvalid'));
-    }
-    if (!name) {
-      throw new Error(t('settings.azureDeploymentNameRequired'));
-    }
-    if (/\s/u.test(name)) {
-      throw new Error(t('settings.azureDeploymentNameWhitespace'));
-    }
-    if (!connectApiKey.trim()) {
-      throw new Error(t('settings.apiKeyRequired'));
-    }
-    const contextLengthRaw = connectContextLength.trim();
-    let contextLength: number | undefined;
-    if (contextLengthRaw) {
-      const parsed = parseModelContextLength(Number(contextLengthRaw));
-      if (parsed === undefined) {
-        throw new Error(t('settings.contextLengthInvalid'));
-      }
-      contextLength = parsed;
-    }
-    await onAddModel({
-      groupId: defaultPresetProviderGroupId("azure"),
-      name,
-      apiBase: azureApiBaseFromResourceName(azureResourceName),
-      apiKey: connectApiKey,
-      provider: "azure",
-      transportKind: "open-responses",
-      azureResourceName,
-      ...(contextLength !== undefined ? { contextLength } : {}),
-    });
-    setConnectDialogOpen(false);
-    runAfterRadixOverlayClose(resetConnectWizard);
-  };
-
-  const saveCloudflareSingle = async () => {
-    if (selectedProvider !== "cloudflare-ai-gateway") {
-      return;
-    }
-    const accountId = connectCloudflareAccountId.trim();
-    const gatewayId = connectCloudflareGatewayId.trim();
-    const name = connectName.trim();
-    if (!accountId) {
-      throw new Error(t('settings.cloudflareAccountIdRequired'));
-    }
-    if (!isValidCloudflareAccountId(accountId)) {
-      throw new Error(t('settings.cloudflareAccountIdInvalid'));
-    }
-    if (!gatewayId) {
-      throw new Error(t('settings.cloudflareGatewayIdRequired'));
-    }
-    if (!isValidCloudflareGatewayId(gatewayId)) {
-      throw new Error(t('settings.cloudflareGatewayIdInvalid'));
-    }
-    if (!name) {
-      throw new Error(t('settings.modelNameRequired'));
-    }
-    if (!connectApiKey.trim()) {
-      throw new Error(t('settings.cloudflareAiGatewayApiTokenRequired'));
-    }
-    const contextLengthRaw = connectContextLength.trim();
-    let contextLength: number | undefined;
-    if (contextLengthRaw) {
-      const parsed = parseModelContextLength(Number(contextLengthRaw));
-      if (parsed === undefined) {
-        throw new Error(t('settings.contextLengthInvalid'));
-      }
-      contextLength = parsed;
-    }
-    await onAddModel({
-      groupId: defaultPresetProviderGroupId("cloudflare-ai-gateway"),
-      name,
-      apiBase: cloudflareAiGatewayApiBaseFromAccountId(accountId),
-      apiKey: connectApiKey,
-      provider: "cloudflare-ai-gateway",
-      transportKind: connectTransportKindForRequest ?? defaultConnectTransportKind("cloudflare-ai-gateway"),
-      cloudflareAccountId: accountId,
-      cloudflareGatewayId: gatewayId,
-      ...(contextLength !== undefined ? { contextLength } : {}),
-    });
-    setConnectDialogOpen(false);
-    runAfterRadixOverlayClose(resetConnectWizard);
-  };
-
-  const saveVertexExpressSingle = async () => {
-    if (selectedProvider !== "google-vertex-ai" || vertexConnectMode !== "express") {
-      return;
-    }
-    const name = connectName.trim();
-    if (!name) {
-      throw new Error(t('settings.modelNameRequired'));
-    }
-    if (!connectApiKey.trim()) {
-      throw new Error(t('settings.apiKeyRequired'));
-    }
-    const contextLengthRaw = connectContextLength.trim();
-    let contextLength: number | undefined;
-    if (contextLengthRaw) {
-      const parsed = parseModelContextLength(Number(contextLengthRaw));
-      if (parsed === undefined) {
-        throw new Error(t('settings.contextLengthInvalid'));
-      }
-      contextLength = parsed;
-    }
-    await onAddModel({
-      groupId: defaultPresetProviderGroupId("google-vertex-ai"),
-      name,
-      apiBase: "",
-      apiKey: connectApiKey,
-      provider: "google-vertex-ai",
-      transportKind: "openai-compatible",
-      ...(contextLength !== undefined ? { contextLength } : {}),
-    });
-    setConnectDialogOpen(false);
-    runAfterRadixOverlayClose(resetConnectWizard);
   };
 
   return (
@@ -1468,15 +928,7 @@ export function ModelsSettingsPanel({
                   <p className="px-2 py-6 text-center text-sm text-muted-foreground">{t('app.noMatches')}</p>
                 ) : (
                   filteredProviders.map((row) => (
-                    <button
-                      key={row.id}
-                      type="button"
-                      className="flex w-full items-center gap-2.5 rounded-md px-3 py-2.5 text-left text-sm hover:bg-muted/60 focus-visible:bg-muted/60 focus-visible:outline-none"
-                      onClick={() => startConnect(row.id)}
-                    >
-                      <ProviderIcon providerId={row.id} />
-                      <span className="min-w-0 flex-1 truncate">{row.label}</span>
-                    </button>
+                    <ProviderPickerRowButton key={row.id} row={row} onSelect={startConnect} />
                   ))
                 )}
               </div>
@@ -1485,852 +937,19 @@ export function ModelsSettingsPanel({
         </DialogContent>
       </Dialog>
 
-      <Dialog
-        open={connectDialogOpen}
-        // 不在 onOpenChange(false) 立刻 resetConnectWizard：Dialog exit 动画未结束时 Content 仍挂载，
-        // selectedProvider 被清 null 会导致标题回退「连接提供商」、条件 Select 消失。见 runAfterRadixOverlayClose / openProviderPicker。
-        onOpenChange={setConnectDialogOpen}
-      >
-        <DialogContent className="sm:max-w-lg" showCloseButton>
-          <DialogHeader>
-            <DialogTitle>
-              {selectedProvider === "custom" ? t('settings.customConnection') : selectedProviderLabel}
-            </DialogTitle>
-            <DialogDescription>
-              {selectedProvider === "custom"
-                ? t('settings.customConnectionDescription')
-                : selectedProvider === "volcengine"
-                  ? t('settings.volcengineConnectionDescription')
-                  : selectedProvider === "siliconflow"
-                    ? t('settings.siliconflowConnectionDescription')
-                    : selectedProvider === "moonshot-ai"
-                      ? t('settings.moonshotConnectionDescription')
-                      : selectedProvider === "minimax"
-                        ? t('settings.minimaxConnectionDescription')
-                        : selectedProvider === "alibaba"
-                          ? t('settings.alibabaConnectionDescription')
-                  : providerShowsConnectTransportPicker(selectedProvider)
-                      ? t('settings.providerConnectionDescription')
-                      : t('settings.providerSimpleDescription')}
-            </DialogDescription>
-          </DialogHeader>
-
-          <div className="grid gap-3 py-1">
-            {selectedProvider === "alibaba" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-alibaba-billing-mode">{t('settings.alibabaBillingMode')}</Label>
-                <div className={DESKTOP_FORM_INPUT_SHELL}>
-                  <Select
-                    value={connectAlibabaBillingMode}
-                    onValueChange={(value) =>
-                      setConnectAlibabaBillingMode(value as "standard" | "token-plan")
-                    }
-                  >
-                    <SelectTrigger
-                      id="connect-alibaba-billing-mode"
-                      className={DESKTOP_FORM_FIELD_TRIGGER_INNER}
-                    >
-                      <SelectValue />
-                    </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value="standard">
-                        {t('settings.alibabaBillingModeStandard')}
-                      </SelectItem>
-                      <SelectItem value="token-plan">
-                        {t('settings.alibabaBillingModeTokenPlan')}
-                      </SelectItem>
-                    </SelectContent>
-                  </Select>
-                </div>
-              </div>
-            ) : null}
-            {selectedProvider === "stepfun" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-stepfun-billing-mode">{t('settings.stepfunBillingMode')}</Label>
-                <div className={DESKTOP_FORM_INPUT_SHELL}>
-                  <Select
-                    value={connectStepfunBillingMode}
-                    onValueChange={(value) =>
-                      setConnectStepfunBillingMode(value as "standard" | "step-plan")
-                    }
-                  >
-                    <SelectTrigger
-                      id="connect-stepfun-billing-mode"
-                      className={DESKTOP_FORM_FIELD_TRIGGER_INNER}
-                    >
-                      <SelectValue />
-                    </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value="standard">
-                        {t('settings.stepfunBillingModeStandard')}
-                      </SelectItem>
-                      <SelectItem value="step-plan">
-                        {t('settings.stepfunBillingModeStepPlan')}
-                      </SelectItem>
-                    </SelectContent>
-                  </Select>
-                </div>
-              </div>
-            ) : null}
-            {selectedProvider && providerSupportsSiteSelection(selectedProvider) && !connectAlibabaIsTokenPlan ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-provider-site">{t('settings.site')}</Label>
-                <div className={DESKTOP_FORM_INPUT_SHELL}>
-                  <Select
-                    value={connectProviderSite}
-                    onValueChange={setConnectProviderSite}
-                  >
-                    <SelectTrigger
-                      id="connect-provider-site"
-                      className={DESKTOP_FORM_FIELD_TRIGGER_INNER}
-                    >
-                      <SelectValue />
-                    </SelectTrigger>
-                    <SelectContent>
-                      {listProviderConnectSiteOptions(selectedProvider).map((option) => (
-                        <SelectItem key={option.id} value={option.id}>
-                          {t(option.labelKey, { defaultValue: option.fallbackLabel })}
-                        </SelectItem>
-                      ))}
-                    </SelectContent>
-                  </Select>
-                </div>
-              </div>
-            ) : null}
-            {connectAlibabaWorkspaceRequired ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-alibaba-workspace-id">{t('settings.alibabaWorkspaceId')}</Label>
-                <DesktopFormInput
-                  id="connect-alibaba-workspace-id"
-                  value={connectAlibabaWorkspaceId}
-                  onChange={(e) => setConnectAlibabaWorkspaceId(e.target.value)}
-                  placeholder={t('settings.alibabaWorkspaceIdPlaceholder')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {selectedProvider && providerShowsConnectTransportPicker(selectedProvider) ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-api-transport">{t('settings.apiType')}</Label>
-                <div className={DESKTOP_FORM_INPUT_SHELL}>
-                  <Select
-                    value={connectTransportKind}
-                    onValueChange={(value) => setConnectTransportKind(value as DesktopTransportKind)}
-                  >
-                    <SelectTrigger
-                      id="connect-api-transport"
-                      className={DESKTOP_FORM_FIELD_TRIGGER_INNER}
-                    >
-                      <SelectValue />
-                    </SelectTrigger>
-                  <SelectContent>
-                    {connectTransportOptionsForProvider(selectedProvider).map((option) => (
-                      <SelectItem key={`${option.value}-${option.label}`} value={option.value}>
-                        {option.label}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                  </Select>
-                </div>
-                {(() => {
-                  const transportSummary = connectTransportOptionsForProvider(selectedProvider)
-                    .filter((option) => option.value === connectTransportKind)
-                    .map((option) => connectTransportOptionSummary(option, selectedProvider))
-                    .find(Boolean);
-                  return transportSummary ? (
-                    <p className="text-xs leading-5 text-muted-foreground">{transportSummary}</p>
-                  ) : null;
-                })()}
-              </div>
-            ) : null}
-            {selectedProvider === "custom" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-custom-group-label">{t('settings.customGroupLabel')}</Label>
-                <DesktopFormInput
-                  id="connect-custom-group-label"
-                  value={connectCustomGroupLabel}
-                  onChange={(e) => setConnectCustomGroupLabel(e.target.value)}
-                  placeholder={t('settings.customGroupLabelPlaceholder')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "custom" ? (
-              <div className="grid gap-2">
-                <Label>{t('settings.modelAddMode')}</Label>
-                <div
-                  role="tablist"
-                  aria-label={t('settings.modelAddMode')}
-                  className="inline-flex h-9 shrink-0 rounded-lg border border-border/40 bg-muted/30 p-0.5"
-                >
-                  {(["single", "bulk"] as const).map((value) => (
-                    <button
-                      key={value}
-                      type="button"
-                      role="tab"
-                      aria-selected={customConnectMode === value}
-                      className={cn(
-                        DESKTOP_EDITOR_TAB_CLASS,
-                        customConnectMode === value
-                          ? "bg-background text-foreground shadow-sm"
-                          : "text-muted-foreground hover:text-foreground",
-                      )}
-                      disabled={modelsBusy || modelsPreviewBusy}
-                      onClick={() => setCustomConnectMode(value)}
-                    >
-                      {value === "single" ? t('settings.addSingle') : t('settings.addAll')}
-                    </button>
-                  ))}
-                </div>
-              </div>
-            ) : null}
-            {selectedProvider === "custom" && customConnectMode === "single" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-model-name">{t('settings.modelName')}</Label>
-                <DesktopFormInput
-                  id="connect-model-name"
-                  value={connectName}
-                  onChange={(e) => setConnectName(e.target.value)}
-                  placeholder={t('settings.modelNameExample')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "custom" && customConnectMode === "single" ? (
-              <div className="grid gap-2">
-                <Label>{t('settings.modelCapabilities')}</Label>
-                <ModelCapabilitiesCombobox
-                  value={connectCapabilities}
-                  disabled={modelsBusy || modelsPreviewBusy}
-                  onChange={setConnectCapabilities}
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "custom" && customConnectMode === "single" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-context-length">{t('settings.contextLength')}</Label>
-                <DesktopFormInput
-                  id="connect-context-length"
-                  type="number"
-                  min={1}
-                  step={1}
-                  value={connectContextLength}
-                  onChange={(e) => setConnectContextLength(e.target.value)}
-                  placeholder={t('settings.optional')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "custom" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-api-base">{t('settings.endpoint')}</Label>
-                <DesktopFormInput
-                  id="connect-api-base"
-                  value={connectApiBase}
-                  onChange={(e) => setConnectApiBase(e.target.value)}
-                  placeholder={t('settings.endpointPlaceholder')}
-                  autoComplete="off"
-                  required
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "amazon-bedrock" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-aws-region">{t('settings.bedrockRegion')}</Label>
-                <DesktopFormInput
-                  id="connect-aws-region"
-                  value={connectAwsRegion}
-                  onChange={(e) => setConnectAwsRegion(e.target.value)}
-                  placeholder={t('settings.bedrockRegionPlaceholder')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "amazon-bedrock" ? (
-              <div className="grid gap-2">
-                <Label>{t('settings.bedrockAuthMode')}</Label>
-                <div
-                  role="tablist"
-                  aria-label={t('settings.bedrockAuthMode')}
-                  className="inline-flex h-9 shrink-0 rounded-lg border border-border/40 bg-muted/30 p-0.5"
-                >
-                  {(["bearer", "iam"] as const).map((value) => (
-                    <button
-                      key={value}
-                      type="button"
-                      role="tab"
-                      aria-selected={bedrockConnectMode === value}
-                      className={cn(
-                        DESKTOP_EDITOR_TAB_CLASS,
-                        bedrockConnectMode === value
-                          ? "bg-background text-foreground shadow-sm"
-                          : "text-muted-foreground hover:text-foreground",
-                      )}
-                      disabled={modelsBusy || modelsPreviewBusy}
-                      onClick={() => setBedrockConnectMode(value)}
-                    >
-                      {value === "bearer"
-                        ? t('settings.bedrockAuthBearer')
-                        : t('settings.bedrockAuthIam')}
-                    </button>
-                  ))}
-                </div>
-              </div>
-            ) : null}
-            {selectedProvider === "amazon-bedrock" && bedrockConnectMode === "bearer" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-bedrock-model-name">{t('settings.modelName')}</Label>
-                <DesktopFormInput
-                  id="connect-bedrock-model-name"
-                  value={connectName}
-                  onChange={(e) => setConnectName(e.target.value)}
-                  placeholder={t('settings.bedrockModelNamePlaceholder')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "amazon-bedrock" && bedrockConnectMode === "bearer" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-context-length-bedrock">{t('settings.contextLength')}</Label>
-                <DesktopFormInput
-                  id="connect-context-length-bedrock"
-                  type="number"
-                  min={1}
-                  step={1}
-                  value={connectContextLength}
-                  onChange={(e) => setConnectContextLength(e.target.value)}
-                  placeholder={t('settings.optional')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "azure" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-azure-resource-name">{t('settings.azureResourceName')}</Label>
-                <DesktopFormInput
-                  id="connect-azure-resource-name"
-                  value={connectAzureResourceName}
-                  onChange={(e) => setConnectAzureResourceName(e.target.value)}
-                  placeholder={t('settings.azureResourceNamePlaceholder')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "azure" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-azure-deployment-name">{t('settings.azureDeploymentName')}</Label>
-                <DesktopFormInput
-                  id="connect-azure-deployment-name"
-                  value={connectName}
-                  onChange={(e) => setConnectName(e.target.value)}
-                  placeholder={t('settings.azureDeploymentNamePlaceholder')}
-                  autoComplete="off"
-                />
-                <p className="text-xs leading-5 text-muted-foreground">
-                  {t('settings.azureSingleHint')}
-                </p>
-              </div>
-            ) : null}
-            {selectedProvider === "azure" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-context-length-azure">{t('settings.contextLength')}</Label>
-                <DesktopFormInput
-                  id="connect-context-length-azure"
-                  type="number"
-                  min={1}
-                  step={1}
-                  value={connectContextLength}
-                  onChange={(e) => setConnectContextLength(e.target.value)}
-                  placeholder={t('settings.optional')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "cloudflare-ai-gateway" ? (
-              <>
-                <div className="grid gap-2">
-                  <Label htmlFor="connect-cloudflare-account-id">{t('settings.cloudflareAccountId')}</Label>
-                  <DesktopFormInput
-                    id="connect-cloudflare-account-id"
-                    value={connectCloudflareAccountId}
-                    onChange={(e) => setConnectCloudflareAccountId(e.target.value)}
-                    placeholder={t('settings.cloudflareAccountIdPlaceholder')}
-                    autoComplete="off"
-                  />
-                </div>
-                <div className="grid gap-2">
-                  <Label htmlFor="connect-cloudflare-gateway-id">{t('settings.cloudflareGatewayId')}</Label>
-                  <DesktopFormInput
-                    id="connect-cloudflare-gateway-id"
-                    value={connectCloudflareGatewayId}
-                    onChange={(e) => setConnectCloudflareGatewayId(e.target.value)}
-                    placeholder={t('settings.cloudflareGatewayIdPlaceholder')}
-                    autoComplete="off"
-                  />
-                </div>
-              </>
-            ) : null}
-            {selectedProvider === "cloudflare-ai-gateway" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-cloudflare-model-name">{t('settings.modelName')}</Label>
-                <DesktopFormInput
-                  id="connect-cloudflare-model-name"
-                  value={connectName}
-                  onChange={(e) => setConnectName(e.target.value)}
-                  placeholder={t('settings.cloudflareModelNamePlaceholder')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "cloudflare-ai-gateway" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-context-length-cloudflare">{t('settings.contextLength')}</Label>
-                <DesktopFormInput
-                  id="connect-context-length-cloudflare"
-                  type="number"
-                  min={1}
-                  step={1}
-                  value={connectContextLength}
-                  onChange={(e) => setConnectContextLength(e.target.value)}
-                  placeholder={t('settings.optional')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {(selectedProvider !== "amazon-bedrock" || bedrockConnectMode === "bearer")
-            && selectedProvider !== "google-vertex-ai" ? (
-            <div className="grid gap-2">
-              <Label htmlFor="connect-api-key">
-                {selectedProvider === "amazon-bedrock"
-                  ? t('settings.bedrockApiKeyLabel')
-                  : selectedProvider === "cloudflare-ai-gateway"
-                    ? t('settings.cloudflareAiGatewayApiToken')
-                  : "API Key"}
-              </Label>
-              <DesktopFormInput
-                id="connect-api-key"
-                type="password"
-                value={connectApiKey}
-                onChange={(e) => setConnectApiKey(e.target.value)}
-                placeholder={
-                  selectedProvider === "cloudflare-ai-gateway"
-                    ? t('settings.cloudflareAiGatewayApiTokenPlaceholder')
-                    : t('settings.enterApiKey')
-                }
-                autoComplete="off"
-              />
-              {selectedProvider === "amazon-bedrock" ? (
-                <p className="text-xs leading-5 text-muted-foreground">
-                  {t('settings.bedrockBearerHint')}
-                </p>
-              ) : null}
-            </div>
-            ) : null}
-            {selectedProvider === "amazon-bedrock" && bedrockConnectMode === "iam" ? (
-              <>
-                <div className="grid gap-2">
-                  <Label htmlFor="connect-access-key-id">{t('settings.bedrockAccessKeyId')}</Label>
-                  <DesktopFormInput
-                    id="connect-access-key-id"
-                    value={connectAccessKeyId}
-                    onChange={(e) => setConnectAccessKeyId(e.target.value)}
-                    placeholder={t('settings.enterApiKey')}
-                    autoComplete="off"
-                  />
-                </div>
-                <div className="grid gap-2">
-                  <Label htmlFor="connect-secret-access-key">{t('settings.bedrockSecretAccessKey')}</Label>
-                  <DesktopFormInput
-                    id="connect-secret-access-key"
-                    type="password"
-                    value={connectSecretAccessKey}
-                    onChange={(e) => setConnectSecretAccessKey(e.target.value)}
-                    placeholder={t('settings.enterApiKey')}
-                    autoComplete="off"
-                  />
-                </div>
-                <p className="text-xs leading-5 text-muted-foreground">
-                  {t('settings.bedrockIamHint')}
-                </p>
-              </>
-            ) : null}
-            {selectedProvider === "google-vertex-ai" ? (
-              <div className="grid gap-2">
-                <Label>{t('settings.vertexAuthMode')}</Label>
-                <div
-                  role="tablist"
-                  aria-label={t('settings.vertexAuthMode')}
-                  className="inline-flex h-9 shrink-0 rounded-lg border border-border/40 bg-muted/30 p-0.5"
-                >
-                  {(["adc", "service-account", "express"] as const).map((value) => (
-                    <button
-                      key={value}
-                      type="button"
-                      role="tab"
-                      aria-selected={vertexConnectMode === value}
-                      className={cn(
-                        DESKTOP_EDITOR_TAB_CLASS,
-                        vertexConnectMode === value
-                          ? "bg-background text-foreground shadow-sm"
-                          : "text-muted-foreground hover:text-foreground",
-                      )}
-                      disabled={modelsBusy || modelsPreviewBusy}
-                      onClick={() => setVertexConnectMode(value)}
-                    >
-                      {value === "adc"
-                        ? t('settings.vertexAuthAdc')
-                        : value === "service-account"
-                          ? t('settings.vertexAuthServiceAccount')
-                          : t('settings.vertexAuthExpress')}
-                    </button>
-                  ))}
-                </div>
-              </div>
-            ) : null}
-            {selectedProvider === "google-vertex-ai" && vertexConnectMode === "express" ? (
-              <div className="grid gap-2">
-                <Label htmlFor="connect-vertex-api-key">API Key</Label>
-                <DesktopFormInput
-                  id="connect-vertex-api-key"
-                  type="password"
-                  value={connectApiKey}
-                  onChange={(e) => setConnectApiKey(e.target.value)}
-                  placeholder={t('settings.enterApiKey')}
-                  autoComplete="off"
-                />
-              </div>
-            ) : null}
-            {selectedProvider === "google-vertex-ai" && vertexConnectMode !== "express" ? (
-              <>
-                <div className="grid gap-2">
-                  <Label htmlFor="connect-vertex-project">{t('settings.vertexProject')}</Label>
-                  <DesktopFormInput
-                    id="connect-vertex-project"
-                    value={connectVertexProject}
-                    onChange={(e) => setConnectVertexProject(e.target.value)}
-                    placeholder={t('settings.vertexProjectPlaceholder')}
-                    autoComplete="off"
-                  />
-                </div>
-                <div className="grid gap-2">
-                  <Label htmlFor="connect-vertex-location">{t('settings.vertexLocation')}</Label>
-                  <DesktopFormInput
-                    id="connect-vertex-location"
-                    value={connectVertexLocation}
-                    onChange={(e) => setConnectVertexLocation(e.target.value)}
-                    placeholder={t('settings.vertexLocationPlaceholder')}
-                    autoComplete="off"
-                  />
-                </div>
-              </>
-            ) : null}
-            {selectedProvider === "google-vertex-ai" && vertexConnectMode === "adc" ? (
-              <p className="text-xs leading-5 text-muted-foreground">
-                {t('settings.vertexAdcHint')}
-              </p>
-            ) : null}
-            {selectedProvider === "google-vertex-ai" && vertexConnectMode === "express" ? (
-              <>
-                <div className="grid gap-2">
-                  <Label htmlFor="connect-vertex-model-name">{t('settings.modelName')}</Label>
-                  <DesktopFormInput
-                    id="connect-vertex-model-name"
-                    value={connectName}
-                    onChange={(e) => setConnectName(e.target.value)}
-                    placeholder={t('settings.vertexModelNamePlaceholder')}
-                    autoComplete="off"
-                  />
-                </div>
-                <div className="grid gap-2">
-                  <Label htmlFor="connect-context-length-vertex">{t('settings.contextLength')}</Label>
-                  <DesktopFormInput
-                    id="connect-context-length-vertex"
-                    type="number"
-                    min={1}
-                    step={1}
-                    value={connectContextLength}
-                    onChange={(e) => setConnectContextLength(e.target.value)}
-                    placeholder={t('settings.optional')}
-                    autoComplete="off"
-                  />
-                </div>
-              </>
-            ) : null}
-            {selectedProvider === "google-vertex-ai" && vertexConnectMode === "service-account" ? (
-              <>
-                <div className="grid gap-2">
-                  <Label htmlFor="connect-vertex-client-email">{t('settings.vertexClientEmail')}</Label>
-                  <DesktopFormInput
-                    id="connect-vertex-client-email"
-                    value={connectVertexClientEmail}
-                    onChange={(e) => setConnectVertexClientEmail(e.target.value)}
-                    placeholder={t('settings.vertexClientEmailPlaceholder')}
-                    autoComplete="off"
-                  />
-                </div>
-                <div className="grid gap-2">
-                  <Label htmlFor="connect-vertex-private-key">{t('settings.vertexPrivateKey')}</Label>
-                  <DesktopFormInput
-                    id="connect-vertex-private-key"
-                    type="password"
-                    value={connectVertexPrivateKey}
-                    onChange={(e) => setConnectVertexPrivateKey(e.target.value)}
-                    placeholder={t('settings.enterApiKey')}
-                    autoComplete="off"
-                  />
-                </div>
-                <p className="text-xs leading-5 text-muted-foreground">
-                  {t('settings.vertexServiceAccountHint')}
-                </p>
-              </>
-            ) : null}
-          </div>
-          <DialogFooter>
-            <DialogFooterActions>
-              <Button
-                type="button"
-                variant="outline"
-                size="sm"
-                onClick={() => setConnectDialogOpen(false)}
-                disabled={modelsBusy || modelsPreviewBusy}
-              >
-                {t('common.cancel')}
-              </Button>
-                {selectedProvider === "custom" && customConnectMode === "single" ? (
-                  <Button
-                    type="button"
-                    size="sm"
-                    disabled={
-                      modelsBusy ||
-                      modelsPreviewBusy ||
-                      !connectName.trim() ||
-                      !connectApiBase.trim() ||
-                      !connectApiKey.trim()
-                    }
-                    onClick={() => {
-                      void (async () => {
-                        try {
-                          await saveCustomSingle();
-                        } catch {
-                          /* runtimeError */
-                        }
-                      })();
-                    }}
-                  >
-                    {modelsBusy ? <LoaderCircle className="size-4 animate-spin" /> : null}
-                    {t('settings.addThisModel')}
-                  </Button>
-                ) : null}
-                {selectedProvider === "custom" && customConnectMode === "bulk" ? (
-                  <Button
-                    type="button"
-                    size="sm"
-                    disabled={
-                      modelsBusy
-                      || modelsPreviewBusy
-                      || !connectApiBase.trim()
-                      || !connectApiKey.trim()
-                    }
-                    onClick={() => {
-                      void (async () => {
-                        try {
-                          await syncCatalogFromUpstream(true);
-                        } catch {
-                          /* runtimeError */
-                        }
-                      })();
-                    }}
-                  >
-                    {modelsBusy || modelsPreviewBusy ? (
-                      <LoaderCircle className="size-4 animate-spin" />
-                    ) : (
-                      <RefreshCw className="size-4" />
-                    )}
-                    {t('settings.addProvider')}
-                  </Button>
-                ) : null}
-                {selectedProvider === "amazon-bedrock" && bedrockConnectMode === "bearer" ? (
-                  <Button
-                    type="button"
-                    size="sm"
-                    disabled={
-                      modelsBusy
-                      || modelsPreviewBusy
-                      || !connectAwsRegion.trim()
-                      || !connectName.trim()
-                      || !connectApiKey.trim()
-                    }
-                    onClick={() => {
-                      void (async () => {
-                        try {
-                          await saveBedrockBearerSingle();
-                        } catch {
-                          /* runtimeError */
-                        }
-                      })();
-                    }}
-                  >
-                    {modelsBusy ? <LoaderCircle className="size-4 animate-spin" /> : null}
-                    {t('settings.addThisModel')}
-                  </Button>
-                ) : null}
-                {selectedProvider === "amazon-bedrock" && bedrockConnectMode === "iam" ? (
-                  <Button
-                    type="button"
-                    size="sm"
-                    disabled={
-                      modelsBusy
-                      || modelsPreviewBusy
-                      || !connectAwsRegion.trim()
-                      || !hasBedrockCatalogCredentials
-                    }
-                    onClick={() => {
-                      void (async () => {
-                        try {
-                          await syncCatalogFromUpstream(false);
-                        } catch {
-                          /* runtimeError */
-                        }
-                      })();
-                    }}
-                  >
-                    {modelsBusy || modelsPreviewBusy ? (
-                      <LoaderCircle className="size-4 animate-spin" />
-                    ) : null}
-                    {t('settings.addProvider')}
-                  </Button>
-                ) : null}
-                {selectedProvider === "google-vertex-ai" && vertexConnectMode === "express" ? (
-                  <Button
-                    type="button"
-                    size="sm"
-                    disabled={
-                      modelsBusy
-                      || modelsPreviewBusy
-                      || !connectName.trim()
-                      || !connectApiKey.trim()
-                    }
-                    onClick={() => {
-                      void (async () => {
-                        try {
-                          await saveVertexExpressSingle();
-                        } catch {
-                          /* runtimeError */
-                        }
-                      })();
-                    }}
-                  >
-                    {modelsBusy ? <LoaderCircle className="size-4 animate-spin" /> : null}
-                    {t('settings.addThisModel')}
-                  </Button>
-                ) : null}
-                {selectedProvider === "google-vertex-ai" && vertexConnectMode !== "express" ? (
-                  <Button
-                    type="button"
-                    size="sm"
-                    disabled={
-                      modelsBusy
-                      || modelsPreviewBusy
-                      || !connectVertexProject.trim()
-                      || !connectVertexLocation.trim()
-                      || (vertexConnectMode === "service-account" && !hasVertexCatalogCredentials)
-                    }
-                    onClick={() => {
-                      void (async () => {
-                        try {
-                          await syncCatalogFromUpstream(false);
-                        } catch {
-                          /* runtimeError */
-                        }
-                      })();
-                    }}
-                  >
-                    {modelsBusy || modelsPreviewBusy ? (
-                      <LoaderCircle className="size-4 animate-spin" />
-                    ) : null}
-                    {t('settings.addProvider')}
-                  </Button>
-                ) : null}
-                {selectedProvider === "azure" ? (
-                  <Button
-                    type="button"
-                    size="sm"
-                    disabled={
-                      modelsBusy
-                      || modelsPreviewBusy
-                      || !connectAzureResourceName.trim()
-                      || !connectName.trim()
-                      || !connectApiKey.trim()
-                    }
-                    onClick={() => {
-                      void (async () => {
-                        try {
-                          await saveAzureSingle();
-                        } catch {
-                          /* runtimeError */
-                        }
-                      })();
-                    }}
-                  >
-                    {modelsBusy ? <LoaderCircle className="size-4 animate-spin" /> : null}
-                    {t('settings.addThisModel')}
-                  </Button>
-                ) : null}
-                {selectedProvider === "cloudflare-ai-gateway" ? (
-                  <Button
-                    type="button"
-                    size="sm"
-                    disabled={
-                      modelsBusy
-                      || modelsPreviewBusy
-                      || !connectCloudflareAccountId.trim()
-                      || !connectCloudflareGatewayId.trim()
-                      || !connectName.trim()
-                      || !connectApiKey.trim()
-                    }
-                    onClick={() => {
-                      void (async () => {
-                        try {
-                          await saveCloudflareSingle();
-                        } catch {
-                          /* runtimeError */
-                        }
-                      })();
-                    }}
-                  >
-                    {modelsBusy ? <LoaderCircle className="size-4 animate-spin" /> : null}
-                    {t('settings.addThisModel')}
-                  </Button>
-                ) : null}
-                {selectedProvider !== null && selectedProvider !== "custom" && selectedProvider !== "amazon-bedrock" && selectedProvider !== "azure" && selectedProvider !== "google-vertex-ai" && selectedProvider !== "cloudflare-ai-gateway" ? (
-                  <Button
-                    type="button"
-                    size="sm"
-                    disabled={
-                      modelsBusy
-                      || modelsPreviewBusy
-                      || !connectApiKey.trim()
-                    }
-                    onClick={() => {
-                      void (async () => {
-                        try {
-                          await syncCatalogFromUpstream(false);
-                        } catch {
-                          /* runtimeError */
-                        }
-                      })();
-                    }}
-                  >
-                    {modelsBusy || modelsPreviewBusy ? (
-                      <LoaderCircle className="size-4 animate-spin" />
-                    ) : null}
-                    {t('settings.addProvider')}
-                  </Button>
-                ) : null}
-            </DialogFooterActions>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      {connectTarget !== null ? (
+        <ProviderConnectDialog
+          key={`${connectTarget.provider}-${connectTarget.session}`}
+          provider={connectTarget.provider}
+          open={connectDialogOpen}
+          onOpenChange={setConnectDialogOpen}
+          busy={modelsBusy}
+          previewBusy={modelsPreviewBusy}
+          onAddModel={onAddModel}
+          onAddProviderModels={onAddProviderModels}
+          onPreviewModels={onPreviewModels}
+        />
+      ) : null}
     </div>
   );
 }
